Make views require dedicated `unsafe` marker traits. (#32)
* Make views require dedicated `unsafe` marker traits.

Requiring `Into` and `From` conversions is not sufficient for
`FilelikeView` and `SocketlikeView`, becuase there's no guarantee
that the `Into` implementation will return the same handle as the `From`
implementation. If a type allows its handle to be reassigned, that
can lead to the old handle being freed twice, and the new handle being
leaked.

To fix this, introduce unsafe marker traits `FilelikeViewType` and
`SocketlikeViewType`, and have `FilelikeView` and `SocketlikeView`
require these traits.

* Fix io_lifetimes_use_std compilation.

* Remove the `DerefMut` from the view types.

* Add API tests for the `&*` idiom for using `Read` and `Write`.

* Add a debug check for views checking that they return the same handle.

* Use `ManuallyDrop` to avoid needing `Option` in the view types.

`ManuallyDrop::take` allows the view types to properly consume their temporary
objects without needing an `.unwrap()`.