Update for io_safety being stabilized. (#41)
* Update for io_safety being stabilized.

io_safety is now [stable in Rust 1.63]! This PR updates io-lifetimes to
use the standard library types and traits when available, and use its
own types and traits on older Rust versions.

The traits `FromFd` and `IntoFd` are now marked as deprecated. These are
replaced by `From<OwnedFd>` and `From<...> for OwnedFd` in the standard
library, and users should migrate accordingly.

[stable in Rust 1.63]: https://blog.rust-lang.org/2022/08/11/Rust-1.63.0.html#rust-ownership-for-raw-file-descriptorshandles-io-safety

* Disable deprecation warnings for our own uses of `FromFd` and `IntoFd`.

And, migrate the portability types and traits to use `From<OwnedFd>`
and `Into<OwnedFd>`.

* Deprecate the `from_fd` function, rather than the `FromFd` trait.

This allows the `from_into_fd` extension to continue working, for now.